extract Cursor
