+++ lib/graphql/language_server/cursor/input_stack.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+module GraphQL
+  class LanguageServer
+    class Cursor
+      class InputStack < SelfStack
+      end
+    end
+  end
+end
